make sure that binary prefix trees created on 64-bit archs work on 32-bit machines and vice-versa (cannot use size_t since it is defined to have a different size on the two platforms)
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@914 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/File.h
@@ -26,26 +26,26 @@ template<typename T> inline void fRead(FILE* f,T& t)  {
 }
 
 template<typename T> inline size_t fWrite(FILE* f,const T* b,const T* e) {
-  size_t s=std::distance(b,e);size_t rv=fWrite(f,s);
+  uint32_t s=std::distance(b,e);size_t rv=fWrite(f,s);
   if(fwrite(b,sizeof(T),s,f)!=s) {std::cerr<<"ERROR: fwrite!\n";abort();}
   return rv+sizeof(T)*s;
 }
 
 template<typename T> inline size_t fWrite(FILE* f,const T b,const T e) {
-  size_t s=std::distance(b,e);size_t rv=fWrite(f,s);
+  uint32_t s=std::distance(b,e);size_t rv=fWrite(f,s);
   if(fwrite(&(*b),sizeof(T),s,f)!=s) {std::cerr<<"ERROR: fwrite!\n";abort();}
   return rv+sizeof(T)*s;
 }
 
 template<typename C> inline size_t fWriteVector(FILE* f,const C& v) {
-  size_t s=v.size();
+  uint32_t s=v.size();
   size_t rv=fWrite(f,s);
   if(fwrite(&v[0],sizeof(typename C::value_type),s,f)!=s) {std::cerr<<"ERROR: fwrite!\n";abort();}
   return rv+sizeof(typename C::value_type)*s;
 }
 
 template<typename C> inline void fReadVector(FILE* f, C& v) {
-  size_t s;fRead(f,s);v.resize(s);
+  uint32_t s;fRead(f,s);v.resize(s);
   size_t r=fread(&(*v.begin()),sizeof(typename C::value_type),s,f);
   if(r!=s) {
     std::cerr<<"ERROR: freadVec! "<<r<<" "<<s<<"\n";abort();}
